CPU: Fix single-step when branch mispredict is blocked on StageF
When trying to single-step over a branch misprediction, sometimes StageF
is busy, causing the instruction to stall in Stage1. However,
stage1_halted does not reflect this, and so we halt rl_pipe at the end
of this cycle, causing the instruction to not be committed and no step
to have taken place, even though rl_stage1_stop is guarded by StageF not
being busy. Instead, keep the pipeline running until StageF is ready to
be redirect, much like we do for the downstream stages.