MMU_Cache: Drop rg_state predicates from do_req
This list of predicates is non-exhaustive, missing MODULE_RUNNING and
several of the special RSP states. However, by having these predicates,
we force ourselves to conflict with rl_probe_and_immed_rsp, and so once
we are in the MODULE_RUNNING state, we will discard all future requests,
as scheduling ends up treating rl_probe_and_immed_rsp as more urgent.
So, we go back to something that looks more like what was there before,
namely that we don't check rg_state.

If this hurts timing (or the scheduler warnings are annoying),
mutually_exclusive can be used to assert that nobody should be making
requests when in the busy states.