Don't hard-code arbitrary path to bsc-contrib
Install your bsc-contrib into the same PREFIX as bsc. There is no other
good way to tell bsc where bsc-contrib is. It is not the responsibility
of every Makefile in the world to track this. If and when bsc learns to
have multiple BLUESPECDIR's, that can be configured by the user, and
this Makefile should continue to just work. But until then, separate
bsc-contrib installations are not supported, and have very little value.
It is no effort to just install bsc-contrib to the same place and keep
the world free of these hacks.