TLB: More possible timing improvements
The exc_code, pte, pte_modified and allow_cap fields should be ignored
when we have a TLB miss, and all but exc_code should be ignored if we
trap. Thus don't needlessly condition them on whether we got a TLB hit
(but be careful not to affect the disabled translation case).